% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_repair_cost.R
\name{bridge_repair_cost}
\alias{bridge_repair_cost}
\title{Cost of Bridge Repairs due to flooding}
\usage{
bridge_repair_cost(floods, costperflood, discount, time)
}
\arguments{
\item{floods}{number of flooding events per year based on data}

\item{costperflood}{estimated cost of repairs needed per flooding event ($USD) default = $1 million}

\item{discount}{discount rate (no unit) default = 7%}

\item{time}{number of years evaluation default = 30 years}
}
\value{
npv (USD) net present value
}
\description{
This function calculates the total cost of bridge repairs based on the extent of local flooding.
}
