% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode function}
\usage{
mode(v)
}
\arguments{
\item{v}{IntegerVector}
}
\value{
integer vector of length 1 with mode or length 0
}
\description{
Task:
Write a function mode() to determine the most frequently occurring value in an integer
vector (mode). If the mode is ambiguous (e.g. for 1, 2, 2, 2, 3, 3, 1, 3), return any mode
}

