% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{permutations function}
\usage{
perms(n)
}
\arguments{
\item{n}{integer}
}
\value{
matrix with n! rows and n columns with permutations
}
\description{
Task:
Write an Rcpp function perms() to generate all the possible permutations of the set {1, 2, . . . , n}
(for some given n). The function should return a matrix with n! rows and n columns
}

