% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qgramdist}
\alias{qgramdist}
\title{qgramdist}
\usage{
qgramdist(x, y, q)
}
\arguments{
\item{x}{a vector of single (<10 if q>1) integers}

\item{y}{a vector of single (<10 if q>1) integers}

\item{q}{an integer, q in q-gram distance}
}
\value{
an integer value - q-gram distance, Inf if q is larger then a vector's length
}
\description{
Calculates a q-gram distance \url{https://journal.r-project.org/archive/2014-1/loo.pdf} 
between two vectors of integers.
}

