% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{simplifies list of numeric vectors to array if possibles}
\usage{
simplify2array(L)
}
\arguments{
\item{L}{list of numeric vectors}
}
\value{
vector, array or list
}
\description{
Task:
Write your own implementation of the simplify2array() function. Your function should
expect a list of numeric vectors on input.
}

