% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{a vector of integers}
}
\value{
an integer value - mode
}
\description{
Calculates a mode number, which is the most frequently occuring number in a given vector. 
If the mode is ambigous returns a smaller number of equal mode.
}

